+++ xskillscore/core/accessor.py
@@ -2,6 +2,7 @@ import xarray as xr
 
 from .deterministic import (
     effective_sample_size,
+    linslope,
     mae,
     mape,
     me,
@@ -45,6 +46,11 @@ class XSkillScoreAccessor(object):
             return x
         else:
             return self._obj[x]
+
+    def linslope(self, a, b, *args, **kwargs):
+        a = self._in_ds(a)
+        b = self._in_ds(b)
+        return linslope(a, b, *args, **kwargs)
 
     def pearson_r(self, a, b, *args, **kwargs):
         a = self._in_ds(a)